conncache: fix compiler warning
warning C4267: '=' : conversion from 'size_t' to 'long', possible loss
of data

The member connection_id of struct connectdata is a long (always a
32-bit signed integer on Visual C++) and the member next_connection_id
of struct conncache is a size_t, so one of them should be changed to
match the other.

This patch the size_t in struct conncache to long (the less invasive
change as that variable is only ever used in a single code line).

Bug: http://curl.haxx.se/bug/view.cgi?id=1399

+++ lib/conncache.h
@@ -25,7 +25,7 @@
 struct conncache {
   struct curl_hash *hash;
   size_t num_connections;
-  size_t next_connection_id;
+  long next_connection_id;
 };
 
 struct conncache *Curl_conncache_init(int size);
